/main/login -> /로 변경 .gif 위치 변경 import 형식은 찾아보니까 jsx 파일에서 쓰는 방식이고 css 파일에서는 경로 지정으로 하는 것같더라구요. src 파일 만들어서 그쪽으로 경로만 변경했습니다.
