+++ pages/index.tsx
@@ -1,10 +1,29 @@
-import type {NextPage} from 'next';
-import Head from 'next/head';
-import Image from 'next/image';
-import styles from '../styles/Home.module.css';
+import classes from './index.module.css';
 
-const Home: NextPage = () => {
-    return <div>Code Formatting</div>;
-};
+function LogIn(props) {
+    return (
+        <div className={classes.bbody}>
+            <div className={classes.body}></div>
+            <div className={classes.grad}></div>
+            <div className={classes.header}>
+                <div>
+                    성프야<span>Fantasy</span>
+                </div>
+            </div>
+            <br></br>
+            <div className={classes.login}>
+                <input type='text' placeholder='username' name='user'></input>
+                <br></br>
+                <input
+                    type='password'
+                    placeholder='password'
+                    name='password'
+                ></input>
+                <br></br>
+                <input type='button' value='Login'></input>
+            </div>
+        </div>
+    );
+}
 
-export default Home;
+export default LogIn;
